+++ src/models/mongo/poi.js
@@ -7,6 +7,7 @@ const poiSchema = new Schema({
   description : String,
   latitude    : Number,
   longitude   : Number,
+  image       : String,
   categoryID: {
     type: Schema.Types.ObjectId,
     ref: "Category",